% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardev-package.r
\docType{package}
\name{standardev}
\alias{standardev}
\alias{standardev-package}
\title{standardev: a standard deviation calculator}
\description{
A simple standard deviation calculator to help user calculate standard deviation
of a given numeric/integer vector
}

